requirements.txt
streamlit==1.32.0
groq==0.5.0
pymupdf==1.24.2
sentence-transformers==2.7.0
numpy==1.26.4
python-dotenv==1.0.1
faiss-cpu==1.8.0
